space_train_requirements.txt
unsloth
trl
peft
datasets
pyyaml
gradio
huggingface_hub
weave
torch
transformers
accelerate
bitsandbytes
